feat: Create and test Pouch class
A Pouch class was created to hold the potion ingredients of the
witch.
This class works with getters and setters.
Methods included in this class can display, add and remove ingredients from
the pouch.
The class was thoroughly tested.
Importing issues between the Pouch and Ingredient class have been resolved.