feat: Create classes Pet, Dog and Cat
The abstract class Pet was created that contains the properties of a pet
like name, age, color, sex with pronoun and hunger status.
This class works with getters and setters.
This class enables the player to adopt a pet, call their pet,
pet, feed and play with their pet as well as check on their
hunger status.
The pet starts out hungry till it gets food, and gets hungry again
after playing.
The Pet class is the parent class of the classes Dog and Cat.

Dog is the child class of Pet and works
with getters and setters.
It has its own dog art.
The dog can be looked at and give paw.

Cat is the child class of Pet and works
with getters and setters.
It has its own cat art.
The cat can be looked at and purrs
when petted.