feat: Create and test Spell class
This class represents all spells and works with getters and
setters.
The class contains the name of the spell, its description,
the kind of effect and its effectiveness as well as its cost
when cast.
Furthermore it is possible to retrieve a description of the
spell's effect.
The class has been thoroughly tested.