feat: Create and test Cell class
This class represents a single cell in the grid of the forest.
It contains information about its coordinates in the grid, its
level, what the witch will encounter there, whether the witch
is currently there and a list of descriptions for all cells.
This class works with getters and setters.
It is possible to enter and to leave the cell as well as to
get a description of the cell's environment.
The class has been thoroughly tested.