feat: Create and test Forest class
This class represents all forests of the game.
It contains information about the grid layout (e.g. 3x3), the
current position of the witch on the forest grid and a list
of the cells of this forest's grid.
This class works with getters and setters.
It allows the witch to enter and leave a cell.
The class has been thoroughly tested.

The witch is now able to move inside the forest.
She can't move outside the grid.
When she enters a new cell, she receives a
description of the environment.