feat: Create introduction, player and forest
An introduction to the game has been added.
If the player is willing to play, a new Player
is created.
If the player wants to start a new game,
the forest for the first level is created.

To do that a static method has been added
to the Forest class, to create the necessary
level 1 grid for the forest.